backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["yyysolhhh <solbaram37@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = ">=4.2,<5.0"
django-environ = "^0.11.2"
djangorestframework = ">=3.14,<4.0"
django-debug-toolbar = "^4.3.0"
django-extensions = "^3.2.3"
mysqlclient = "^2.2.4"
drf-spectacular = "^0.27.2"
dj-rest-auth = "^5.1.0"
djangorestframework-simplejwt = "^5.3.1"
django-allauth = "^0.61.1"
boto3 = "^1.34.76"
pillow = "^10.3.0"
django-storages = "^1.14.2"
djangorestframework-stubs = "^3.14.5"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
isort = "^5.13.2"
mypy = "^1.9.0"
django-stubs = "^4.2.7"
ipython = "^8.23.0"
django-stubs-ext = "^4.2.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
python_version = 3.12
strict = true

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "manage"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "mysite.settings"
